Add GameAnalytics library for ios under modules/
PLEASE note the modifications I had to do in order for the library to
compile and actually function correctly. I had to convert some Strings
in GameAnalytics.cpp to const char *arrays and it wasn't exactly
trivial because the Godot core engine defines CharType as wchar_t. Just
search for "jeyu" in GameAnalytics.cpp for the full comment.